Move make_search_scope() to outdated.py.
