+++ src/pip/_internal/commands/list.py
@@ -7,7 +7,6 @@ from pip._vendor import six
 from pip._vendor.six.moves import zip_longest
 
 from pip._internal.cli import cmdoptions
-from pip._internal.cli.cmdoptions import make_search_scope
 from pip._internal.cli.req_command import IndexGroupCommand
 from pip._internal.exceptions import CommandError
 from pip._internal.index import PackageFinder
@@ -17,6 +16,7 @@ from pip._internal.utils.misc import (
     get_installed_distributions,
     write_output,
 )
+from pip._internal.utils.outdated import make_search_scope
 from pip._internal.utils.packaging import get_installer
 
 logger = logging.getLogger(__name__)
